Add files via upload
INSIGHTS :

The average price in London boroughs is higher compared to the rest of England. It has been affected by major financial and political events (such the Recession and Brexit), but has significantly increased from 1995 to 2019 (by a factor of 5),
Affluent regions such as Kensighton and Westiminister have the highest average price,
The number of houses sold plummeted after the recession, streadily increased until 2016 but then again dropped after the referendum,
As anyone could predict, the main factor that influences the average price is the financial prosperity of the corresponding borough. Higher salaries and more jobs result in higher prices.